Cleanup (#131)
As @emberian correctly points out in https://github.com/jwilger/eventcore/discussions/129, the code in this project prior to this change was full of a lot of LLM slop that, at the time, I was hoping I would get to cleaning up quickly, but then my attention was required elsewhere, and this project just sat for about three months. In the meantime, my process for integrating LLM-augmented coding into my workflow has changed significantly for the better. At this point, the best path forward is to scrap everything that was here and take another run at it.